+++ assets/src/apps/app.tsx
@@ -23,7 +23,7 @@ export function defineApplication<T extends State>(Component: React.FunctionComp
       content: parsedContent,
     };
 
-    // console.log('MOUNTED', { mountPoint, params, props });
+    // console.log('MOUNT UP', props);
 
     ReactDOM.render(
       <Provider store={store}>