+++ lib/oli_web/live/curriculum/drop_target.ex
@@ -7,7 +7,8 @@ defmodule OliWeb.Curriculum.DropTarget do
 
   def render(assigns) do
     ~L"""
-    <div phx-hook="DropTarget" id="<%= @id %>"
+    <div phx-hook="DropTarget"
+      id="drop-target-<%= assigns.index %>"
       data-drop-index="<%= assigns.index %>"
       style="height: 15px;"
     >