+++ lib/oli_web/views/view_helpers.ex
@@ -10,8 +10,10 @@ defmodule OliWeb.ViewHelpers do
   alias Oli.Branding.Brand
 
   def brand_logo_html(conn_or_brand, opts \\ [])
+
   def brand_logo_html(%Brand{name: name, logo: logo, logo_dark: logo_dark}, opts) do
     class = Keyword.get(opts, :class, "")
+
     ~E"""
       <img src="<%= logo %>" height="40" class="d-dark-none <%= class %>" alt="<%= name %>">
       <img src="<%= value_or(logo_dark, logo) %>" height="40" class="d-light-none <%= class %>"  alt="<%= name %>">
@@ -21,6 +23,7 @@ defmodule OliWeb.ViewHelpers do
   def brand_logo_html(conn, opts) do
     class = Keyword.get(opts, :class, "")
     section = conn.assigns[:section]
+
     ~E"""
       <img src="<%= brand_logo_url(section) %>" height="40" class="d-dark-none <%= class %>" alt="<%= brand_name(section) %>">
       <img src="<%= brand_logo_url_dark(section) %>" height="40" class="d-light-none <%= class %>"  alt="<%= brand_name(section) %>">
@@ -104,5 +107,4 @@ defmodule OliWeb.ViewHelpers do
         end
     end
   end
-
 end