+++ priv/repo/migrations/20210421162709_create_part_component_registrations.exs
@@ -16,7 +16,9 @@ defmodule Oli.Repo.Migrations.CreatePartComponentRegistrations do
       timestamps(type: :timestamptz)
     end
 
-    create unique_index(:part_component_registrations, [:slug], name: :index_part_component_slug_registrations)
+    create unique_index(:part_component_registrations, [:slug],
+             name: :index_part_component_slug_registrations
+           )
 
     create unique_index(:part_component_registrations, [:delivery_element],
              name: :index_part_component_delivery_element_registrations
@@ -36,14 +38,19 @@ defmodule Oli.Repo.Migrations.CreatePartComponentRegistrations do
 
     create table(:part_component_registration_projects, primary_key: false) do
       timestamps(type: :timestamptz)
-      add :part_component_registration_id, references(:part_component_registrations), primary_key: true
+
+      add :part_component_registration_id, references(:part_component_registrations),
+        primary_key: true
+
       add :project_id, references(:projects), primary_key: true
     end
 
     create index(:part_component_registration_projects, [:part_component_registration_id])
     create index(:part_component_registration_projects, [:project_id])
 
-    create unique_index(:part_component_registration_projects, [:part_component_registration_id, :project_id],
+    create unique_index(
+             :part_component_registration_projects,
+             [:part_component_registration_id, :project_id],
              name: :index_part_component_registration_project
            )
   end